requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Deprecated==1.2.13
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
numpy==1.22.2
openreview-py==1.1.1
pandas==1.4.1
pycryptodome==3.14.1
PyJWT==2.3.0
pylatexenc==2.10
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
tld==0.10
tqdm==4.62.3
urllib3==1.26.8
visitor==0.1.3
Werkzeug==2.0.3
wrapt==1.13.3
WTForms==3.0.1
